feat: 演示模式支持显示一些密钥类的配置但会从安全层面做限制
