+++ app/views/admin/settings/index.html.slim
@@ -34,7 +34,7 @@ form.form-horizontal
 
                 display_value = case value
                                 when Hash
-                                  JSON.pretty_generate(value)
+                                  pretty_json(value)
                                 when Array
                                   value.join(', ')
                                 when TrueClass
@@ -47,7 +47,10 @@ form.form-horizontal
               dl.system-info
                 dt = t("admin.settings.#{key}")
                 dd
-                  - if params[:readonly]
+                  - if params[:readonly] || (value.is_a?(Hash) && secure_key?(value))
                     pre.disabled = display_value
                   - else
                     pre = link_to display_value, edit_admin_setting_path(key)
+
+                  - if key == 'demo_mode'
+                    small = t('admin.settings.demo_mode_tips')
