+++ app/views/admin/system_info/index.html.slim
@@ -110,7 +110,7 @@
             dt = key
             dd
               pre
-                = value
+                = secure_value(key, value)
 
     .card.collapsed-card
       .card-header
