+++ config/locales/zealot/zh-CN.yml
@@ -65,14 +65,15 @@ zh-CN:
       site_locale: 网站语言
       site_domain: 网站域名
 
-      presets: 默认值
+      presets: 预设值
       default_schemes: 应用类型模板
-      default_role: 初始化权限
+      default_role: 新用户权限
 
       switch_mode: 模式开关
       registrations_mode: 注册模式
-      guest_mode: 开启游客模式
-      demo_mode: 演示模式（定时恢复默认数据）
+      guest_mode: 游客模式
+      demo_mode: 演示模式
+      demo_mode_tips: 开启演示模式会触发：每天定时恢复默认数据；无法编辑或删除默认管理员；隐私数据会安全化显示且无法编辑
 
       third_party_auth: 第三方登录
       ldap: LDAP
@@ -80,8 +81,8 @@ zh-CN:
       gitlab: Gitlab
       google_oauth: Google OAuth
 
-      backup: 备份
-      misc: 杂项
+      # backup: 备份
+      # misc: 杂项
 
       stmp: 邮件配置
       mailer_default_from: 默认邮件发件地址
